+++ lib/src/core/controllers/renderer_questionnaire_controller.dart
@@ -1,4 +1,5 @@
 import 'package:fhir_r4/fhir_r4.dart';
+import 'package:fhir_renderer_questionnaire/src/core/utils/fhir_renderer_questionnaire_response_utils.dart';
 import 'package:flutter/material.dart';
 
 import '../data/field_behavioral_data.dart';
@@ -15,6 +16,12 @@ class RendererQuestionnaireController {
   ///
   /// This is typically set by the `BaseQuestionnaireRenderer` when the renderer is initialized.
   QuestionnaireResponse Function()? onGenerateQuestionnaireResponse;
+
+  /// The FHIR Questionnaire to be rendered.
+  final Questionnaire questionnaire;
+
+  /// The initial questionnaire response.
+  QuestionnaireResponse? initialQuestionnaireResponse;
 
   /// Optional [PageController] for controlling page navigation in [QuestionnairePageViewRenderer].
   final PageController? pageViewController;
@@ -36,18 +43,38 @@ class RendererQuestionnaireController {
   final Map<String, FieldBehavioralData> indexedItems = {};
 
   /// Creates a [RendererQuestionnaireController].
+  ///
+  /// The [questionnaire] parameter is required and defines the structure of the
+  /// questionnaire to be rendered.
+  ///
+  /// The [initialQuestionnaireResponse] can be provided to pre-fill the questionnaire.
+  /// If not provided, an initial response will be generated based on the [questionnaire].
+  ///
+  /// [sliversInitialIndex] determines the initial scroll position when using
+  /// [QuestionnaireSliversViewRenderer].
+  ///
+  /// [listViewScrollController] and [pageViewController] allow external control
+  /// over scrolling and pagination.
   RendererQuestionnaireController({
+    required this.questionnaire,
     this.sliversInitialIndex = 0,
     this.listViewScrollController,
     this.pageViewController,
+    this.initialQuestionnaireResponse,
     this.onGenerateQuestionnaireResponse,
-  });
+  }) {
+    initialQuestionnaireResponse ??= FhirRendererQuestionnaireResponseUtils
+        .generateInitialQuestionnaireResponse(
+      questionnaire,
+    );
+  }
 
   /// Generates the current [QuestionnaireResponse] from the renderer's state.
   ///
   /// This method calls the [onGenerateQuestionnaireResponse] callback which collects
   /// the data from the widget tree and triggers validation of required fields.
   QuestionnaireResponse generateQuestionnaireResponse() {
-    return onGenerateQuestionnaireResponse!();
+    return onGenerateQuestionnaireResponse?.call() ??
+        initialQuestionnaireResponse!;
   }
 }